Restructure hooks
